Rewrite event binding to be more specific
I've rewritten the initial event binding to use $('desc').parent()
instead of checking every element for a desc child. This should
improve performance, although I haven't benchmarked it.

I've also pulled the event bindings into a public function in lieu
of using $.delegate or $.on. This is not as nice as using one of
the jQuery deferred functions, but since there is not a CSS selector
equivalent of $('desc').parent(), and since there is no convenient
guaranteed-present parent element outside of document or svg
(neither of which is great to use performance-wise), and since
the use case of dynamic SVG elements currently seems to be rare,
I think this is a reasonable way of accomplishing the event binding
while still having the possibility of rebinding on element insertion.